requirements.txt.txt
flask==2.0.1 
google-cloud-aiplatform==1.59.0 
google-cloud-discoveryengine==0.11.0 
gunicorn==20.1.0